campaign-service: chuẩn hoá /api/v1/*, đổi Integer→Long, fix Feign configs

+++ campaign-service/src/main/java/com/oem/evwarranty/dto/request/AffectedVehicleUpdateRequest.java
@@ -3,15 +3,15 @@ package com.oem.evwarranty.dto.request;
 import com.oem.evwarranty.model.enums.AffectedStatus;
 
 public class AffectedVehicleUpdateRequest {
-    private Integer assignedServiceCenterId;
+    private Long assignedServiceCenterId;
     private AffectedStatus status; // PENDING/NOTIFIED/SCHEDULED/COMPLETED ...
     // getter/setter ...
 
-    public Integer getAssignedServiceCenterId() {
+    public Long getAssignedServiceCenterId() {
         return assignedServiceCenterId;
     }
 
-    public void setAssignedServiceCenterId(Integer assignedServiceCenterId) {
+    public void setAssignedServiceCenterId(Long assignedServiceCenterId) {
         this.assignedServiceCenterId = assignedServiceCenterId;
     }
