+++ campaign-service/src/main/java/com/oem/evwarranty/dto/response/NotificationResponse.java
@@ -3,19 +3,19 @@ package com.oem.evwarranty.dto.response;
 import java.time.LocalDateTime;
 
 public class NotificationResponse {
-    private Integer notificationId;
-    private Integer campaignId;
-    private Integer affectedId;
+    private Long notificationId;
+    private Long campaignId;
+    private Long affectedId;
     private String channel;
     private String status;
     private LocalDateTime sentAt;
 
-    public Integer getNotificationId() { return notificationId; }
-    public void setNotificationId(Integer notificationId) { this.notificationId = notificationId; }
-    public Integer getCampaignId() { return campaignId; }
-    public void setCampaignId(Integer campaignId) { this.campaignId = campaignId; }
-    public Integer getAffectedId() { return affectedId; }
-    public void setAffectedId(Integer affectedId) { this.affectedId = affectedId; }
+    public Long getNotificationId() { return notificationId; }
+    public void setNotificationId(Long notificationId) { this.notificationId = notificationId; }
+    public Long getCampaignId() { return campaignId; }
+    public void setCampaignId(Long campaignId) { this.campaignId = campaignId; }
+    public Long getAffectedId() { return affectedId; }
+    public void setAffectedId(Long affectedId) { this.affectedId = affectedId; }
     public String getChannel() { return channel; }
     public void setChannel(String channel) { this.channel = channel; }
     public String getStatus() { return status; }
